requirements/app.txt
beautifulsoup4
tabulate
pymongo
sqlalchemy
psycopg2-binary
python-dotenv
requests
pandas
numpy
logging
prefect==2.20.16
azure-storage-blob